Move query_context.hpp from main/ to common/
QueryContext is a tiny header-only value type wrapping
optional_ptr<ClientContext> (which it only forward-declares). It has no real
dependency on main/ and is used by-value in low-level common/ and storage/
headers (file writers, compressed_file_system, block_manager, storage_info),
forcing a layering back-edge that a forward declaration can't remove.

Relocating it to common/ - the layer it actually belongs to - removes the
back-edge at the source for every includer. Pure relocation: no TU's include
closure changes, so total_tu_include_cost is unchanged; the win is layering.

Removes 4 common -> main and 2 storage -> main header back-edges.

common/ header -> main/ edges
before: 18
after:  14
