+++ src/include/duckdb/common/csv_writer.hpp
@@ -11,7 +11,7 @@
 #include "duckdb/common/common.hpp"
 #include "duckdb/execution/operator/csv_scanner/csv_reader_options.hpp"
 #include "duckdb/common/serializer/memory_stream.hpp"
-#include "duckdb/main/query_context.hpp"
+#include "duckdb/common/query_context.hpp"
 
 namespace duckdb {
 class MemoryStream;
