+++ src/include/duckdb/common/query_context.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// duckdb/main/query_context.hpp
+// duckdb/common/query_context.hpp
 //
 //
 //===----------------------------------------------------------------------===//
