+++ src/include/duckdb/common/serializer/buffered_file_writer.hpp
@@ -10,7 +10,7 @@
 
 #include "duckdb/common/serializer/write_stream.hpp"
 #include "duckdb/common/file_system.hpp"
-#include "duckdb/main/query_context.hpp"
+#include "duckdb/common/query_context.hpp"
 
 namespace duckdb {
