+++ src/include/duckdb/main/client_context.hpp
@@ -30,7 +30,7 @@
 #include "duckdb/main/table_description.hpp"
 #include "duckdb/planner/expression/bound_parameter_data.hpp"
 #include "duckdb/transaction/transaction_context.hpp"
-#include "duckdb/main/query_context.hpp"
+#include "duckdb/common/query_context.hpp"
 #include "duckdb/main/query_parameters.hpp"
 
 namespace duckdb {
